Renamed cluster.py to zookeeper.py
